Projects/ma_deeds_scraper/requirements.txt
playwright==1.46.0
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.4
